+++ lua/plugins/orgmode.lua
@@ -2,8 +2,8 @@ return {
    "nvim-orgmode/orgmode",
    ft = { "org" },
    opts = {
-      org_agenda_files = "~/orgfiles/**/*",
-      org_default_notes_file = "~/orgfiles/refile.org",
+      org_agenda_files = "~/notes/org/**/*",
+      org_default_notes_file = "~/notes/org/refile.org",
    },
    -- NOTE: If you are using nvim-treesitter with ~ensure_installed = "all"~
    -- option, add ~org~ to ignore_install